Introduce extension API to ShellSentry
This introduces a new extension API to `ShellSentry` that allows for consumers to bring custom implementations of their own checks to the existing config-defined issues. This allows more advanced and dynamic use-cases, such as an AI-powered log processor.

The general design is described in the docs in `ShellSentryExtension.kt`.

Some other things along the way include
- Adding a `min_confidence` threshold to `ShellSentryConfig` for these checks, allowing them to indicate their confidence and more easily tune whether or not to use their results.
- Rename `IssueThrowable` to `KnownIssue` for clarity and extract a base Throwable class for the shared fillInStackTrace logic.